requirements.txt
langchain==0.3.*
langchain-community==0.3.*
faiss-cpu>=1.8.0
python-dotenv>=1.0.1
pypdf>=4.2.0
docx2txt>=0.8
tiktoken>=0.7.0
langchain-openai>=0.2.0